- New Query on jpa repository - New exception handle for response

+++ src/main/java/com/binar/springboot/crud_rest_task/models/Film.java
@@ -13,7 +13,7 @@ public class Film {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "film_id")
-    private Long filmId;
+    private int filmId;
 
     private String title;
